refactor(dev): process and window filtering functionality in new filters module
