+++ dev/tools/src/lib.rs
@@ -1,0 +1,4 @@
+//! Tools utils
+
+/// Filters for matching windows and processes
+pub mod filters;
